refactor(server): move index.js and app.js to /server
