--- server/.eslintrc.js
@@ -1,9 +1,0 @@
-module.exports = {
-  env: {
-    node: true
-  },
-  plugins: [
-  ],
-  rules: {
-  }
-}
