+++ test/integration/challenges.js
@@ -1,6 +1,6 @@
 const test = require('ava')
 const request = require('supertest')
-const app = require('../../dist/app')
+const app = require('../../dist/server/app')
 const { v4: uuidv4 } = require('uuid')
 const auth = require('../../dist/server/auth')
 const config = require('../../dist/config/server')
