+++ test/integration/leaderboard.js
@@ -1,6 +1,6 @@
 const test = require('ava')
 const request = require('supertest')
-const app = require('../../dist/app')
+const app = require('../../dist/server/app')
 
 const { responseList } = require('../../dist/server/responses')
